Fix impossible atx headings to use setext headings
This changes several cases where ATX headings—which are formed on a
single line—could not be generated safely because of a descendant
which included a line break that could not be escaped.
Headings with a rank of 1 or 2, which include such a line break, are
now generates in setext style—even without an explicit `setext: true`.

Closes GH-26.